-- changed dev settings
