+++ ann/extensions/dataset.py
@@ -27,15 +27,16 @@ from abc import ABC, abstractmethod
 
 import matplotlib.pyplot as plt
 import numpy as np
-import cv2
+# import cv2
 import pandas as pd
 
 from pathlib import Path
-import os
-import gzip
-import requests
+# import os
+# import gzip
+# import requests
 from IPython.display import display
 
+from ann import tensor
 from tensor import Tensor
