+++ ann/utils/validation.py
@@ -1,0 +1,7 @@
+'''
+This modules implements functionalities to validate attribute restrictions liek types and paths
+'''
+
+from typing import Any, Union
+
+# -- under development
